+++ go/descriptors.go
@@ -4,21 +4,21 @@ import "github.com/futzu/bitter"
 
 // SegCmpt Segmentation Descriptor Component
 type SegCmpt struct {
-	ComponentTag uint64
+	ComponentTag uint8
 	PtsOffset    float64
 }
 
 // SpDscptr Splice Descriptor
 type SpDscptr struct {
-	Tag    uint64
-	Length uint64
+	Tag    uint8
+	Length uint8
 	Name   string
 	// identiﬁer 32 uimsbf == 0x43554549 (ASCII “CUEI”)
 	ID string
 
 	ProviderAvailID uint64
 
-	PreRoll                          uint64   `json:",omitempty"`
+	PreRoll                          uint8    `json:",omitempty"`
 	DTMFCount                        uint64   `json:",omitempty"`
 	DTMFChars                        []string `json:",omitempty"`
 	TAISeconds                       uint64   `json:",omitempty"`
@@ -36,11 +36,11 @@ type SpDscptr struct {
 	Components               []SegCmpt `json:",omitempty"`
 	SegmentationDuration     float64   `json:",omitempty"`
 	SegmentationMessage      string    `json:",omitempty"`
-	SegmentationUpidType     uint64    `json:",omitempty"`
+	SegmentationUpidType     uint8     `json:",omitempty"`
 	SegmentationUpidTypeName string    `json:",omitempty"`
-	SegmentationUpidLength   uint64    `json:",omitempty"`
+	SegmentationUpidLength   uint8     `json:",omitempty"`
 	SegmentationUpid         string    `json:",omitempty"`
-	SegmentationTypeID       uint64    `json:",omitempty"`
+	SegmentationTypeID       uint8     `json:",omitempty"`
 	SegmentNum               uint64    `json:",omitempty"`
 	SegmentsExpected         uint64    `json:",omitempty"`
 	SubSegmentNum            uint64    `json:",omitempty"`
@@ -54,8 +54,8 @@ type SpDscptr struct {
 
 // MetaData for splice descriptors
 func (dscptr *SpDscptr) MetaData(bitn *bitter.Bitn) {
-	dscptr.Tag = bitn.AsUInt64(8)
-	dscptr.Length = bitn.AsUInt64(8)
+	dscptr.Tag = bitn.AsUInt8(8)
+	dscptr.Length = bitn.AsUInt8(8)
 	dscptr.ID = bitn.AsHex(32)
 }
 
@@ -86,13 +86,13 @@ func (dscptr *SpDscptr) AvailDscptr(bitn *bitter.Bitn) {
 // DTMFDscptr DTMF Splice Descriptor
 func (dscptr *SpDscptr) DTMFDscptr(bitn *bitter.Bitn) {
 	dscptr.Name = "DTMF Descriptor"
-	dscptr.PreRoll = bitn.AsUInt64(8)
+	dscptr.PreRoll = bitn.AsUInt8(8)
 	dscptr.DTMFCount = bitn.AsUInt64(3)
 	bitn.Forward(5)
 	var dchars [256]string
 	dscptr.DTMFChars = dchars[0:dscptr.DTMFCount]
 	for i := range dscptr.DTMFChars {
-		dscptr.DTMFChars[i] = string(bitn.AsUInt64(8))
+		dscptr.DTMFChars[i] = string(bitn.AsUInt8(8))
 	}
 }
 
@@ -134,10 +134,10 @@ func (dscptr *SpDscptr) decodeSegFlags(bitn *bitter.Bitn) {
 }
 
 func (dscptr *SpDscptr) decodeSegCmpnts(bitn *bitter.Bitn) {
-	ccount := bitn.AsUInt64(8)
+	ccount := bitn.AsUInt8(8)
 	for ccount > 0 { // 6 bytes each
 		ccount--
-		ct := bitn.AsUInt64(8)
+		ct := bitn.AsUInt8(8)
 		bitn.Forward(7)
 		po := bitn.As90k(33)
 		dscptr.Components = append(dscptr.Components, SegCmpt{ct, po})
@@ -148,15 +148,15 @@ func (dscptr *SpDscptr) decodeSegmentation(bitn *bitter.Bitn) {
 	if dscptr.SegmentationDurationFlag {
 		dscptr.SegmentationDuration = bitn.As90k(40)
 	}
-	dscptr.SegmentationUpidType = bitn.AsUInt64(8)
-	dscptr.SegmentationUpidLength = bitn.AsUInt64(8)
+	dscptr.SegmentationUpidType = bitn.AsUInt8(8)
+	dscptr.SegmentationUpidLength = bitn.AsUInt8(8)
 	/**
 	        dscptr.SegmentationUpidTypeName, dscptr.SegmentationUpid = UpidDecoder(
 	            bitn, dscptr.SegmentationUpidType, dscptr.SegmentationUpidLength
 	        )
 	**/
 
-	dscptr.SegmentationTypeID = bitn.AsUInt64(8)
+	dscptr.SegmentationTypeID = bitn.AsUInt8(8)
 
 	/**
 	        if dscptr.SegmentationTypeID in Table22.keys(){
